%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  DevEvalProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.DevEvalProduct}
\alias{as.character.DevEvalProduct}
\alias{DevEvalProduct.as.character}
\alias{as.character,DevEvalProduct-method}
\alias{as.character.DevEvalFileProduct}
\alias{as.character}

\title{Gets a character representation of the product}

\description{
  Gets a character representation of the product.
}

\usage{
\method{as.character}{DevEvalProduct}(x, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{DevEvalProduct}}.
}
\keyword{internal}
\keyword{methods}
