%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  DevEvalProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFullname.DevEvalProduct}
\alias{getFullname.DevEvalProduct}
\alias{DevEvalProduct.getFullname}
\alias{getFullname,DevEvalProduct-method}
\alias{getFullname.DevEvalFileProduct}
\alias{DevEvalProduct.getName}
\alias{getName.DevEvalProduct}
\alias{getName,DevEvalProduct-method}

\alias{DevEvalProduct.getTags}
\alias{getTags.DevEvalProduct}
\alias{getTags,DevEvalProduct-method}

\alias{getFullname}
\alias{getName}
\alias{getTags}

\title{Gets the full name, name and tags}

\description{
  Gets the full name, name and tags consisting of a name and tags.
}

\usage{
\method{getFullname}{DevEvalProduct}(this, ...)
\method{getFullname}{DevEvalFileProduct}(this, ...)
\method{getName}{DevEvalProduct}(this, ...)
\method{getTags}{DevEvalProduct}(this, collapse=",", ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} or a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{DevEvalProduct}}.
}
\keyword{internal}
\keyword{methods}
