%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  DevEvalProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getMimeType.DevEvalFileProduct}
\alias{getMimeType.DevEvalFileProduct}
\alias{DevEvalFileProduct.getMimeType}
\alias{getMimeType,DevEvalFileProduct-method}
\alias{DevEvalFileProduct.getMime}
\alias{getMime.DevEvalFileProduct}
\alias{getMime,DevEvalFileProduct-method}

\alias{getMimeType}
\alias{getMime}

\title{Gets the MIME type}

\description{
 Gets the MIME type.
}

\usage{
\method{getMimeType}{DevEvalFileProduct}(this, default="", ...)
}

\arguments{
  \item{default}{The value returned, if the MIME type could not be inferred.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{DevEvalFileProduct}}.
}
\keyword{internal}
\keyword{methods}
