%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  DevEvalProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPathname.DevEvalFileProduct}
\alias{getPathname.DevEvalFileProduct}
\alias{DevEvalFileProduct.getPathname}
\alias{getPathname,DevEvalFileProduct-method}
\alias{DevEvalFileProduct.getFilename}
\alias{getFilename.DevEvalFileProduct}
\alias{getFilename,DevEvalFileProduct-method}

\alias{DevEvalFileProduct.getPath}
\alias{getPath.DevEvalFileProduct}
\alias{getPath,DevEvalFileProduct-method}

\alias{DevEvalFileProduct.getExtension}
\alias{getExtension.DevEvalFileProduct}
\alias{getExtension,DevEvalFileProduct-method}

\alias{getPathname}
\alias{getFilename}
\alias{getPath}
\alias{getExtension}

\title{Gets the (relative) pathname, filename and path}

\description{
  Gets the (relative) pathname, filename and path.
}

\usage{
\method{getPathname}{DevEvalFileProduct}(this, relative=TRUE, ...)
\method{getFilename}{DevEvalFileProduct}(this, ...)
\method{getPath}{DevEvalFileProduct}(this, ...)
\method{getExtension}{DevEvalFileProduct}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{DevEvalFileProduct}}.
}
\keyword{internal}
\keyword{methods}
