%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  DevEvalProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getType.DevEvalProduct}
\alias{getType.DevEvalProduct}
\alias{DevEvalProduct.getType}
\alias{getType,DevEvalProduct-method}
\alias{getType}

\title{Gets the type}

\description{
 Gets the type.
}

\usage{
\method{getType}{DevEvalProduct}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string or \code{\link[base]{NULL}}.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{DevEvalProduct}}.
}
\keyword{internal}
\keyword{methods}
