% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capturePlot.R
\name{print.RecordedPlot}
\alias{print.RecordedPlot}
\title{Automatically replays a recorded plot}
\usage{
\method{print}{RecordedPlot}(x, ...)
}
\arguments{
\item{x}{A recorded plot of class \code{recordedplot}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
This is identical to the \code{\link[grDevices:replayPlot]{print}()}
method available in \pkg{grDevices}, but before print it also tries
to coerce the architecture internal data structure to that of the
current machine such that it is possible to, for instance, replay a
plot generated on a 32-bit machine on a 64-bit machine.
}
\seealso{
Internally, \code{\link{as.architecture}()} is used
to coerce to the current architecture.
}
\keyword{internal}

