%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  DevEvalProduct.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getDataURI.DevEvalFileProduct}
\alias{getDataURI.DevEvalFileProduct}
\alias{DevEvalFileProduct.getDataURI}
\alias{getDataURI,DevEvalFileProduct-method}
\alias{getDataURI}

\title{Gets a Base64-encoded data URI}

\description{
 Gets a Base64-encoded data URI.
}

\usage{
\method{getDataURI}{DevEvalFileProduct}(this, mime=getMimeType(this), ...)
}

\arguments{
  \item{mime}{The MIME type to be embedded in the data URI.}
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{DevEvalFileProduct}}.
}
\keyword{internal}
\keyword{methods}
