% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capturePlot.R
\name{print.RecordedPlot}
\alias{print.RecordedPlot}
\title{Automatically replays a recorded plot}
\usage{
\method{print}{RecordedPlot}(x, ...)
}
\arguments{
\item{x}{A recorded plot of class \code{recordedplot}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
This is identical to the \code{\link[grDevices:replayPlot]{print}()}
method available in \pkg{grDevices}, but if replaying the plot gives
an error it will also try to replay it after coercing the data structure
to match the architecture of the current machine.  This will make it
possible to, for instance, replay a plot generated on a 32-bit machine
on a 64-bit machine.
}
\seealso{
Internally, \code{\link{as.architecture}()} is used
to coerce to the current architecture.
}
\keyword{internal}
