%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{R.devices-package}
\alias{R.devices-package}
\alias{R.devices}
\docType{package}

\title{Package R.devices}


\description{
  This package provides utility methods that enhance the existing graphical device functions already available in R for the purpose of simplifying the creation of figure files of various image formats and much more.
}

\section{To get started}{
  \itemize{
   \item See the package vignettes.
   \item \code{\link{toEPS}}(), \code{\link{toPDF}}(), \code{\link{toPNG}}(), ... - evaluate
         graphics code and atomically save plot to a figure file.
  }
}

\section{How to cite this package}{
  To cite this package, please use:
\preformatted{
@Manual{,
  title = {R.devices: Enhanced methods for handling graphical devices},
  author = {Henrik Bengtsson},
  year = {2013},
  note = {R package version 2.2.0},
}
}
}

\author{
 Henrik Bengtsson.
}

\section{License}{
 LGPL (>= 2.1).
}
\keyword{package}
